% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{adjust_oag_internal}
\alias{adjust_oag_internal}
\title{Internal OAG Adjustment}
\usage{
adjust_oag_internal(data, from_oag, to_oag, country, ref_year)
}
\arguments{
\item{data}{Input data}

\item{from_oag}{Current OAG}

\item{to_oag}{Target OAG}

\item{country}{Country code}

\item{ref_year}{Reference year}
}
\value{
Data with adjusted OAG
}
\description{
Adjusts the open age group of population data.
}
