% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_tabset}
\alias{app_tabset}
\title{Generate a tabset for the application}
\usage{
app_tabset()
}
\value{
A tabset UI component for the application
}
\description{
Generate a tabset for the application
}
