% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_forecast.R
\name{begin_forecast}
\alias{begin_forecast}
\title{Run Forecast}
\usage{
begin_forecast(
  reactive_pop,
  reactive_tfr,
  reactive_e0,
  reactive_mig,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output,
  simulation_results,
  i18n
)
}
\arguments{
\item{reactive_pop}{Population data}

\item{reactive_tfr}{TFR data}

\item{reactive_e0}{e0 data}

\item{reactive_mig}{migration data}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameter for `\{shiny\}`.}

\item{simulation_results}{A reactive value to store simulation results.}

\item{i18n}{The internationalization object for translations.}
}
\description{
This function manages the simulation process and updates the UI accordingly.
}
