% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{convert_age_groups_internal}
\alias{convert_age_groups_internal}
\title{Internal Age Group Conversion}
\usage{
convert_age_groups_internal(data, from, to, method, country, ref_year)
}
\arguments{
\item{data}{Input data}

\item{from}{Source age type}

\item{to}{Target age type}

\item{method}{Interpolation method}

\item{country}{Country code}

\item{ref_year}{Reference year}
}
\value{
Converted data
}
\description{
Converts between single ages and 5-year groups.
}
