% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_age_group_plot}
\alias{create_age_group_plot}
\title{Create Age Group Plot}
\usage{
create_age_group_plot(pop_dt)
}
\arguments{
\item{dt}{Data table with population data}
}
\value{
A ggplot2 object
}
\description{
Create Age Group Plot
}
