% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{create_age_labels}
\alias{create_age_labels}
\title{Create Age Labels}
\usage{
create_age_labels(age_type, oag)
}
\arguments{
\item{age_type}{"Single Ages" or "5-Year Groups"}

\item{oag}{Open age group value}
}
\value{
Character vector of age labels
}
\description{
Generates appropriate age labels based on age type and OAG.
}
