% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_all_report_plots}
\alias{create_all_report_plots}
\title{Create All Report Plots}
\usage{
create_all_report_plots(
  simulation_results,
  country,
  start_year,
  end_year,
  pop_year = start_year + 1,
  age_group = NULL,
  i18n = NULL
)
}
\arguments{
\item{simulation_results}{List containing all forecast results data}

\item{country}{String containing the country name}

\item{start_year}{Numeric value for the projection start year}

\item{end_year}{Numeric value for the projection end year}

\item{pop_year}{Numeric value for the population pyramid year to display}

\item{age_group}{String indicating the age group for population over time plot}

\item{i18n}{The internationalization object}
}
\value{
A named list containing all generated plots
}
\description{
This function generates non-reactive versions of all plots that appear in the forecast page.
It takes the necessary inputs and creates each plot, then returns them as a named list.
}
