% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_annual_growth_plot}
\alias{create_annual_growth_plot}
\title{Create Annual Growth Rate Time Plot by Age}
\usage{
create_annual_growth_plot(dt, end_year)
}
\arguments{
\item{dt}{Data table with annual growth data.}

\item{end_year}{the date in YYYY-MM-DD where the projection should end.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table to create an annual growth rate time plot by age.
}
