% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_deaths_births_plot}
\alias{create_deaths_births_plot}
\title{Plot births/deaths data with median and confidence intervals}
\usage{
create_deaths_births_plot(
  forecast_birth,
  forecast_death,
  data_type,
  value_type,
  end_year,
  country,
  i18n
)
}
\arguments{
\item{forecast_birth}{A data.table of birth data.}

\item{forecast_death}{A data.table of death data.}

\item{data_type}{A character string indicating the type of data to plot ("births" or "deaths").}

\item{value_type}{A character string indicating the type of value to plot ("count" or "rate").}

\item{end_year}{An integer specifying the upper year limit for the data to be plotted.}

\item{country}{The country for which the data is plotted}

\item{i18n}{The internationalization object for translations.}
}
\value{
A ggplot object.
}
\description{
This function plots the number of births or deaths along with median and confidence intervals
using the ggplot2 package. Data is filtered up to a specified end year.
}
