% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_e0_plot}
\alias{create_e0_plot}
\title{Create e0 Plot}
\usage{
create_e0_plot(dt, end_year, country, i18n = NULL)
}
\arguments{
\item{dt}{Data table with life expectancy data.}

\item{end_year}{the date in YYYY-MM-DD where the projection should end.}

\item{country}{The country for which the data is plotted}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table to create a total life expectancy plot.
}
