% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_e0_projected_plot}
\alias{create_e0_projected_plot}
\title{Create Projected e0 Plot}
\usage{
create_e0_projected_plot(dt, input_sex, country, i18n)
}
\arguments{
\item{dt}{Data table with projected life expectancy data.}

\item{input_sex}{the sex to filter on the e0 projected as a string.}

\item{country}{The country for which the data is plotted}

\item{i18n}{The internationalization object for translations.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table to create a total life expectancy plot from
the model's prediction.
}
