% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{create_enhanced_population_modal_ui}
\alias{create_enhanced_population_modal_ui}
\title{Create Enhanced Population Modal UI}
\usage{
create_enhanced_population_modal_ui(
  modal_id,
  header_title,
  output_id,
  file_input_id,
  download_button_id,
  hide_button_id,
  ref_year = NULL,
  additional_header = NULL,
  i18n = NULL
)
}
\arguments{
\item{modal_id}{A string specifying the ID of the modal.}

\item{header_title}{A string containing the title text for the modal header.}

\item{output_id}{A string specifying the ID of the DT output to be displayed in the modal.}

\item{file_input_id}{A string specifying the ID for the fileInput element.}

\item{download_button_id}{A string specifying the ID for the downloadButton element.}

\item{hide_button_id}{A string specifying the ID for the actionButton used to close the modal.}

\item{ref_year}{(Optional) The reference year to use as default for the date input.}

\item{additional_header}{(Optional) An additional UI element to be included in the header.}

\item{i18n}{The internationalization object.}
}
\value{
A modal UI element with enhanced functionality.
}
\description{
Creates an enhanced modal specifically for population data that supports both
UN WPP data and custom data input with transformation capabilities.
}
