% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{create_header_content}
\alias{create_header_content}
\title{Create Header Content for Modals in Shiny}
\usage{
create_header_content(text, additional_text = NULL, additional_style = "")
}
\arguments{
\item{text}{A string for the primary text content of the header.}

\item{additional_text}{(Optional) Additional text to be included in the header. Default is NULL.}

\item{additional_style}{(Optional) A string specifying CSS styles to be applied to the header. Default is an empty string.}
}
\value{
A `div` element representing the header content.
}
\description{
Constructs a header content `div` for use with create_modal_ui, allowing for
additional text and custom styling.
}
