% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_mig_plot}
\alias{create_mig_plot}
\title{Create Migration Plot}
\usage{
create_mig_plot(dt, end_year, country, i18n)
}
\arguments{
\item{dt}{Data table with migration data.}

\item{end_year}{The date in YYYY-MM-DD where the projection should end.}

\item{country}{The country for which the data is plotted.}

\item{i18n}{The internationalization object for translations.}
}
\value{
A list containing ggplot2 and plotly objects.
}
\description{
This function takes a data table to create a net migration plot.
}
