% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_mig_projected_plot}
\alias{create_mig_projected_plot}
\title{Create Projected Migration Plot}
\usage{
create_mig_projected_plot(dt, end_year, country)
}
\arguments{
\item{dt}{Data table with projected migration data.}

\item{end_year}{The date in YYYY-MM-DD where the projection should end.}

\item{country}{The country for which the data is plotted.}
}
\value{
A list containing ggplot2 and plotly objects.
}
\description{
This function takes a data table to create a projected net migration plot.
}
