% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{create_modal_ui}
\alias{create_modal_ui}
\title{Create a Custom Modal UI for Shiny}
\usage{
create_modal_ui(
  modal_id,
  header_title,
  output_id,
  file_input_id,
  download_button_id,
  hide_button_id,
  additional_header = NULL
)
}
\arguments{
\item{modal_id}{A string specifying the ID of the modal.}

\item{header_title}{A string containing the title text for the modal header.}

\item{output_id}{A string specifying the ID of the DT output to be displayed in the modal.}

\item{file_input_id}{A string specifying the ID for the fileInput element.}

\item{download_button_id}{A string specifying the ID for the downloadButton element.}

\item{hide_button_id}{A string specifying the ID for the actionButton used to close the modal.}

\item{additional_header}{(Optional) An additional UI element to be included in the header.
Default is NULL.}
}
\value{
A modal UI element.
}
\description{
Generates a modal UI element for customizing the TFR and Population data.
This function provides a flexible way to create modals with various elements like
file inputs, download buttons, and custom headers.
}
