% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_pop_aging_pop_size_plot}
\alias{create_pop_aging_pop_size_plot}
\title{Create an Interactive Population Aging and Size Plot}
\usage{
create_pop_aging_pop_size_plot(dt, end_year)
}
\arguments{
\item{dt}{A data frame containing the columns 'year', 'pop', 'percent65',
'un_pop_median', and 'un_percent65'.}

\item{end_year}{A numeric value specifying the last year to include in the plot.}
}
\value{
An interactive plotly object.
}
\description{
This function takes a dataset and an end year to plot population aging and size data
up to the specified year. It plots both projected and UN data for population
and the percentage of the population that is 65 years or older. The plot is interactive,
allowing users to hover over the lines to see data values.
}
