% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_pop_pyramid_plot}
\alias{create_pop_pyramid_plot}
\title{Create Population Pyramid Plot}
\usage{
create_pop_pyramid_plot(dt, country = NULL, input_year = NULL)
}
\arguments{
\item{dt}{Data table with population data.}

\item{country}{The country for which the data is plotted}

\item{input_year}{The input year to filter the data on, default is NULL.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table and an optional input year to create a population pyramid plot.
}
