% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_pop_time_compare_plot}
\alias{create_pop_time_compare_plot}
\title{Create Comparative Pop Time Plot}
\usage{
create_pop_time_compare_plot(dt, input_age, i18n)
}
\arguments{
\item{dt}{Data table containing combined population data with a `simulation` column.}

\item{input_age}{Translated age group selected in the UI.}

\item{i18n}{The internationalization object for translations.}
}
\value{
A named list containing the ggplot object and plotly widget.
}
\description{
Generates a population-over-time comparison plot across multiple simulations.
}
