% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{create_tab}
\alias{create_tab}
\title{Generate a Tab UI Component for the Application}
\usage{
create_tab(tab_name, plot_ui_id, extra_ui = NULL)
}
\arguments{
\item{tab_name}{The name to be displayed on the tab.}

\item{plot_ui_id}{The UI id for the plot.}

\item{extra_ui}{An optional extra UI component.}
}
\value{
A list representing a tab UI component.
}
\description{
Generate a Tab UI Component for the Application
}
