% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_tfr_plot}
\alias{create_tfr_plot}
\title{Create Total Fertility Rate Plot}
\usage{
create_tfr_plot(dt)
}
\arguments{
\item{dt}{Data table with fertility rate data.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table to create a total fertility rate plot.
}
