% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_un_projection_plot}
\alias{create_un_projection_plot}
\title{Create an Interactive Plot Based on Aging and Size Data}
\usage{
create_un_projection_plot(dt, end_year, name_mappings, percent_x = FALSE)
}
\arguments{
\item{dt}{A data frame with a 'year' column and other numeric columns where
some column names contain "un", indicating UN data. The rest are considered
projection data. The function dynamically identifies these columns and does
not assume specific names, making it applicable for various data types.}

\item{end_year}{A numeric value specifying the last year to include in the plot.}

\item{name_mappings}{A named vector where each name corresponds to a column in `dt` that should be renamed, and the value is the new name to be used in the plot. It should also contain a 'title' element for the plot's main title.}

\item{percent_x}{A boolean to whether include a percent label on the X axis}
}
\value{
A list containing a ggplot object and an interactive plotly object.
}
\description{
This function takes a dataset and an end year to plot two sets of data up to
the specified year. One set is identified by the presence of "un" in the column
names, representing UN data, and the other set represents projection data.
The plot is interactive, with hover functionality to display data values.
}
