% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_yadr_oadr_plot}
\alias{create_yadr_oadr_plot}
\title{Plot YADR/OADR data with median and confidence intervals}
\usage{
create_yadr_oadr_plot(oadr, yadr, data_type, end_year, country, i18n)
}
\arguments{
\item{oadr}{A data.table of OADR data.}

\item{yadr}{A data.table of YADR data.}

\item{data_type}{A character string indicating the type of data to plot ("yadr" or "oadr").}

\item{end_year}{An integer specifying the upper year limit for the data to be plotted.}

\item{country}{The country for which the data is plotted}

\item{i18n}{The internationalization object for translations.}
}
\value{
A ggplot object.
}
\description{
This function plots the YADR/OADR along with the UN median and confidence intervals
using the ggplot2 package. Data is filtered up to a specified end year.
}
