% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{A value to test for NULL.}

\item{rhs}{A value to use if lhs is NULL.}
}
\value{
lhs if it is not NULL, otherwise rhs.
}
\description{
This operator returns the left-hand side if it is not NULL,
otherwise it returns the right-hand side.
}
\keyword{internal}
