% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{detect_age_type}
\alias{detect_age_type}
\title{Detect Age Type}
\usage{
detect_age_type(age_vector)
}
\arguments{
\item{age_vector}{Vector of age labels}
}
\value{
"single" or "5-year"
}
\description{
Detects whether age labels are single ages or 5-year groups.
}
