% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{extract_numeric_ages}
\alias{extract_numeric_ages}
\title{Extract Numeric Ages}
\usage{
extract_numeric_ages(age_vector, type)
}
\arguments{
\item{age_vector}{Vector of age labels}

\item{type}{"single" or "5-year"}
}
\value{
Numeric age vector
}
\description{
Extracts numeric age values from age labels.
}
