% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{format_5year_age_labels}
\alias{format_5year_age_labels}
\title{Format 5-Year Age Labels}
\usage{
format_5year_age_labels(ages)
}
\arguments{
\item{ages}{Numeric vector of age starts}
}
\value{
Character vector of formatted age labels
}
\description{
Converts numeric ages from WPP data (0, 5, 10...) to standard
5-year age group labels ("0-4", "5-9"...).
}
