% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{generate_age_labels}
\alias{generate_age_labels}
\title{Generate Age Labels}
\usage{
generate_age_labels(age_type, oag)
}
\arguments{
\item{age_type}{Either "Single Ages" or "5-Year Groups"}

\item{oag}{The open age group value (e.g., 100, 95, 90, etc.)}
}
\value{
Character vector of age labels
}
\description{
Generates age labels for either single year ages or 5-year age groups
with appropriate open age group formatting.
}
