% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_reports.R
\name{generate_demography_report}
\alias{generate_demography_report}
\title{Generate Demographic Projection Report}
\usage{
generate_demography_report(plot_list, country, start_year, end_year, i18n)
}
\arguments{
\item{plot_list}{A named list of ggplot objects generated by `create_all_report_plots`.}

\item{country}{The name of the country for which the report is generated.}

\item{start_year}{The starting year of the projection.}

\item{end_year}{The ending year of the projection.}

\item{i18n}{An i18n translation object from the shiny.i18n package.}
}
\value{
The file path to the generated PDF report.
}
\description{
Creates a PDF report summarizing the demographic projection results, including various plots.
}
