% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_navigation}
\alias{handle_navigation}
\title{Handle Navigation Between Steps}
\usage{
handle_navigation(
  reactive_pop,
  reactive_tfr,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output
)
}
\arguments{
\item{reactive_pop}{A reactive expression that returns the population data to be displayed.}

\item{reactive_tfr}{A reactive expression that returns the TFR data to be displayed.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}`.}
}
\description{
This function manages the navigation between different pages of the application.
}
