% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_pop_tfr_plots}
\alias{handle_pop_tfr_plots}
\title{Render Population and TFR Plots}
\usage{
handle_pop_tfr_plots(
  reactive_pop,
  reactive_tfr,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output
)
}
\arguments{
\item{reactive_pop}{A reactive expression returning the population data.}

\item{reactive_tfr}{A reactive expression returning the TFR (Total Fertility Rate) data.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}.}
}
\description{
Render Population and TFR Plots
}
