% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_report_download}
\alias{handle_report_download}
\title{Handle Report Download}
\usage{
handle_report_download(
  simulation_results,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output,
  i18n = NULL
)
}
\arguments{
\item{simulation_results}{A reactive expression returning the simulation results.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}`.}

\item{i18n}{The internationalization object.}
}
\description{
This function handles the download of a report generated by the application.
It provides a placeholder for actual report generation and download functionality.
}
