% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_validity_checks}
\alias{handle_validity_checks}
\title{Handle all validity checks from the app}
\usage{
handle_validity_checks(
  wpp_starting_year,
  wpp_ending_year,
  input,
  output,
  i18n = NULL,
  has_simulation = NULL
)
}
\arguments{
\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{output}{Internal parameter for `\{shiny\}`.}

\item{i18n}{The internationalization object.}
}
\description{
This function defines all validity checks on inputs to make
sure they work correctly
}
