% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{page_ui}
\alias{page_ui}
\title{Create Page UI Wrapper}
\usage{
page_ui(content)
}
\arguments{
\item{content}{Content to be wrapped in the page UI.}
}
\value{
A shiny UI div for the page layout.
}
\description{
Create Page UI Wrapper
}
