% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{plots_tabset}
\alias{plots_tabset}
\title{Generate the plot modules after simulation}
\usage{
plots_tabset(...)
}
\arguments{
\item{...}{A list of reactive expressions returning ggplot2 objects.}
}
\description{
Generate the plot modules after simulation
}
