% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{prepare_pop_agegroups_table}
\alias{prepare_pop_agegroups_table}
\title{Prepare Population Age Groups Table}
\usage{
prepare_pop_agegroups_table(wpp_dt)
}
\arguments{
\item{wpp_dt}{Data table with population data.}
}
\value{
A shiny.semantic DataTable object.
}
\description{
This function takes a data table to prepare a population age groups table.
}
