% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_e0}
\alias{show_e0}
\title{Show and compute the e0 page}
\usage{
show_e0(reactive_e0, wpp_ending_year, input, output, i18n)
}
\arguments{
\item{reactive_e0}{A reactive function returning the e0 data frame}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}`.}
}
\description{
Show and compute the e0 page
}
