% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_e0_results_ui}
\alias{show_e0_results_ui}
\title{Generate the UI for the e0 page.}
\usage{
show_e0_results_ui()
}
\value{
A div containing UI elements for e0 page
}
\description{
Generate the UI for the e0 page.
}
