% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_forecast_results_ui}
\alias{show_forecast_results_ui}
\title{Generate the UI for the page containing the results of the forecast.}
\usage{
show_forecast_results_ui(input, i18n)
}
\arguments{
\item{input}{Internal parameter for `{shiny}`}
}
\value{
A tabset UI component for the application
}
\description{
Generate the UI for the page containing the results of the forecast.
}
