% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_input_ui}
\alias{show_input_ui}
\title{Generate the UI for the input page.}
\usage{
show_input_ui()
}
\value{
A div containing the input page UI elements.
}
\description{
Generate the UI for the input page.
}
