% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{show_pop_results_ui}
\alias{show_pop_results_ui}
\title{UI component for step one}
\usage{
show_pop_results_ui()
}
\value{
A div containing UI elements for step one
}
\description{
UI component for step one
}
