% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_tfr}
\alias{show_tfr}
\title{Show and compute the TFR page}
\usage{
show_tfr(reactive_tfr, wpp_ending_year, input, output, i18n)
}
\arguments{
\item{reactive_tfr}{A reactive function returning the TFR data frame}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}`.}
}
\description{
Show and compute the TFR page
}
