% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_tfr_results_ui}
\alias{show_tfr_results_ui}
\title{Generate the UI for the TFR page.}
\usage{
show_tfr_results_ui()
}
\value{
A div containing UI elements for TFR page
}
\description{
Generate the UI for the TFR page.
}
