% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{transform_5yr_to_single}
\alias{transform_5yr_to_single}
\title{Transform 5-Year Groups to Single Ages with Standard Parameters}
\usage{
transform_5yr_to_single(data, country, ref_year)
}
\arguments{
\item{data}{Input data frame with 5-year groups}

\item{country}{Country code}

\item{ref_year}{Reference year}
}
\value{
Data frame with single ages, OAG 100
}
\description{
Convenience function for the common transformation from 5-year groups to single ages
using UN method and standard OAG of 100. Used frequently in UN data processing.
}
