% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_transforms.R
\name{transform_population}
\alias{transform_population}
\title{Transform Population}
\usage{
transform_population(
  data,
  from_type,
  to_type,
  from_oag,
  to_oag,
  method,
  country,
  ref_year
)
}
\arguments{
\item{data}{Input data frame}

\item{from_type}{Source age type}

\item{to_type}{Target age type}

\item{from_oag}{Source OAG}

\item{to_oag}{Target OAG}

\item{method}{Interpolation method}

\item{country}{Country code}

\item{ref_year}{Reference year}
}
\value{
Transformed data frame
}
\description{
Core transformation logic for age groups and OAG adjustments.
}
