% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_age_group_plot}
\alias{create_age_group_plot}
\title{Create Age Group Plot}
\usage{
create_age_group_plot(dt, input_scale)
}
\arguments{
\item{dt}{Data table with population data.}

\item{input_scale}{The type of scale to be applied.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table and an input scale to create an age group plot.
}
