% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_pop_pyramid}
\alias{create_pop_pyramid}
\title{Create Population Pyramid Plot}
\usage{
create_pop_pyramid(dt, input_year = NULL)
}
\arguments{
\item{dt}{Data table with population data}
}
\value{
A ggplot2 object
}
\description{
Create Population Pyramid Plot
}
