% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_pop_time_plot}
\alias{create_pop_time_plot}
\title{Create Pop Age Group Time Plot}
\usage{
create_pop_time_plot(dt, input_age)
}
\arguments{
\item{dt}{Data table with population data.}

\item{input_age}{The age group to subset}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table and an input of age group to create a time age group plot.
}
