% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{create_tfr_plot}
\alias{create_tfr_plot}
\title{Create Total Fertility Rate Plot}
\usage{
create_tfr_plot(dt, end_year, country, i18n)
}
\arguments{
\item{dt}{Data table with fertility rate data.}

\item{end_year}{the date in YYYY-MM-DD where the projection should end.}

\item{country}{The country for which the data is plotted}

\item{i18n}{The internationalization object for translations.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a data table to create a total fertility rate plot.
}
