% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_before_analysis_plots}
\alias{handle_before_analysis_plots}
\title{Render Plots Before Analysis}
\usage{
handle_before_analysis_plots(
  reactive_pop,
  reactive_tfr,
  reactive_e0,
  reactive_mig,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output,
  i18n = NULL
)
}
\arguments{
\item{reactive_pop}{A reactive expression returning the population data.}

\item{reactive_tfr}{A reactive expression returning the TFR (Total Fertility Rate) data.}

\item{reactive_e0}{A reactive expression returning the e0 data.}

\item{reactive_mig}{A reactive expression returning the mig data.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}.}

\item{i18n}{The internationalization object.}
}
\description{
Render Plots Before Analysis
}
