% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{handle_customize_data}
\alias{handle_customize_data}
\title{Handle Customization of Data in a Shiny App}
\usage{
handle_customize_data(
  reactive_pop,
  reactive_tfr,
  wpp_starting_year,
  wpp_ending_year,
  input,
  output
)
}
\arguments{
\item{reactive_pop}{A reactive expression that returns the population data to be displayed.}

\item{reactive_tfr}{A reactive expression that returns the TFR data to be displayed.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{input, output}{Internal parameters for `{shiny}`.}
}
\value{
None
}
\description{
This function sets up the UI elements to display and customize population and total fertility rate (TFR) data in a Shiny application. It renders data tables for population and TFR data and sets up modals for data customization.
}
