% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_customize_data}
\alias{handle_customize_data}
\title{Handles Customization of Data in the App}
\usage{
handle_customize_data(
  current_pop_reactive,
  current_tfr_reactive,
  current_e0_reactive,
  current_mig_reactive,
  pop_to_commit_rv,
  tfr_to_commit_rv,
  e0_to_commit_rv,
  mig_to_commit_rv,
  pop_data_source,
  tfr_starting_year,
  wpp_starting_year,
  wpp_ending_year,
  current_tab,
  input,
  output,
  i18n = NULL
)
}
\arguments{
\item{current_pop_reactive}{A reactive expression that returns the population data to be displayed.}

\item{current_tfr_reactive}{A reactive expression that returns the TFR data to be displayed.}

\item{current_e0_reactive}{A reactive expression that returns the e0 data to be displayed.}

\item{current_mig_reactive}{A reactive expression that returns the migration data to be displayed.}

\item{pop_to_commit_rv}{A reactive value to commit the population data.}

\item{tfr_to_commit_rv}{A reactive value to commit the TFR data.}

\item{e0_to_commit_rv}{A reactive value to commit the e0 data.}

\item{mig_to_commit_rv}{A reactive value to commit the migration data.}

\item{tfr_starting_year}{A reactive expression returning the TFR starting year.}

\item{wpp_starting_year}{A reactive expression returning the starting year.}

\item{wpp_ending_year}{A reactive expression returning the ending year.}

\item{current_tab}{A reactive value for the current tab.}

\item{input, output}{Internal parameters for `\{shiny\}`.}

\item{i18n}{The internationalization object.}
}
\value{
None
}
\description{
This function sets up the UI elements to display and customize population and total fertility rate (TFR) data. It renders data tables for population and TFR data and sets up modals for data customization.
}
