% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{location_selector_ui}
\alias{location_selector_ui}
\title{Generate the location selector UI based on the toggle state}
\usage{
location_selector_ui(input, i18n)
}
\arguments{
\item{input}{The input object from the Shiny server function}
}
\value{
A UI element for selecting a country or region
}
\description{
Generate the location selector UI based on the toggle state
}
