% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_show.R
\name{show_pop_results_ui}
\alias{show_pop_results_ui}
\title{Generate the UI for the population page.}
\usage{
show_pop_results_ui()
}
\value{
A div containing UI elements for step one
}
\description{
Generate the UI for the population page.
}
