% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{show_tfr_results_ui}
\alias{show_tfr_results_ui}
\title{UI component for step two}
\usage{
show_tfr_results_ui()
}
\value{
A div containing UI elements for step two
}
\description{
UI component for step two
}
